+++ vast_pipeline/models.py
@@ -205,6 +205,7 @@ class Run(CommentableModel):
     )
     STATUS_CHOICES = [
         ('INI', 'Initialised'),
+        ('QUE', 'Queued'),
         ('RUN', 'Running'),
         ('END', 'Completed'),
         ('ERR', 'Error'),